remove original dom, simplify index files
