+++ src/dom/dom.js
@@ -3,6 +3,8 @@ const belit = require('belit')
 const ninlil = require('ninlil')
 const hyperz = require('hyperz')
 
+const { pushWorkingKeyBranch, popWorkingKeyBranch } = require('../working-key')
+
 /**
  * function to generate a tagged template function for any namespace
  * - for complete documentation, please refer to http://tram-one.io/#tram-dom
@@ -12,13 +14,26 @@ const hyperz = require('hyperz')
  *
  * @return {function}
  */
-const registerDom = (namespace, registry) => {
-  if (registry) {
+const registerDom = (globalSpace = window) => {
+  return (namespace, registry = {}) => {
     assert.equal(typeof registry, 'object', 'Tram-One: registry should be an object')
     assert.ok(!(Array.isArray(registry)), 'Tram-One: registry should be an object')
-  }
 
-  return ninlil(hyperz, belit(namespace), registry || {})
+    // modify the registry so that each component function updates the hook working key
+    const hookedRegistry = Object.keys(registry).reduce((newRegistry, tagName) => {
+      const tagFunction = registry[tagName]
+      const hookedTagFunction = (...args) => {
+        pushWorkingKeyBranch(globalSpace, 'hookKey')(tagName)
+        const tagResult = tagFunction(...args)
+        popWorkingKeyBranch(globalSpace, 'hookKey')()
+        return tagResult
+      }
+
+      return Object.assign({}, newRegistry, {[tagName]: hookedTagFunction})
+    }, {})
+
+    return ninlil(hyperz, belit(namespace), hookedRegistry || {})
+  }
 }
 
 /**
@@ -28,8 +43,8 @@ const registerDom = (namespace, registry) => {
  * @param {object} registry
  * @return {function}
  */
-const registerHtml = (registry) => {
-  return registerDom(null, registry)
+const registerHtml = (globalSpace = window) => (registry) => {
+  return registerDom(globalSpace)(null, registry)
 }
 
 /**
@@ -39,8 +54,8 @@ const registerHtml = (registry) => {
  * @param {object} registry
  * @return {function}
  */
-const registerSvg = (registry) => {
-  return registerDom('http://www.w3.org/2000/svg', registry)
+const registerSvg = (globalSpace = window) => (registry) => {
+  return registerDom(globalSpace)('http://www.w3.org/2000/svg', registry)
 }
 
 module.exports = { registerDom, registerHtml, registerSvg }
