+++ src/engine/index.js
@@ -1,2 +1,1 @@
-const { setupEngine, getEngine, addActions, addListener } = require('./engine')
-module.exports = { setupEngine, getEngine, addActions, addListener }
+module.exports = require('./engine')
