+++ src/log/index.js
@@ -1,2 +1,1 @@
-const { setupLog, getLog, clearLog } = require('./log')
-module.exports = { setupLog, getLog, clearLog }
+module.exports = require('./log')
