+++ src/start/index.js
@@ -1,2 +1,1 @@
-const { start } = require('./start')
-module.exports = { start }
+module.exports = require('./start')
