+++ src/tram-one/index.js
@@ -1,3 +1,1 @@
-const tram = require('./tram-one')
-
-module.exports = tram
+module.exports = require('./tram-one')
